\name{pvmtest-package}
\alias{pvmtest-package}
\alias{pvmtest}
\docType{package}
\title{
\packageTitle{pvmtest}
}
\description{
\packageDescription{pvmtest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pvmtest}
\packageIndices{pvmtest}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pvmtest}

Maintainer: \packageMaintainer{pvmtest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
